third_party/blink/web_tests/svg/dynamic-updates/SVGFECompositeElement-svgdom-in2-prop-expected.txt

SVG 1.1 dynamic update tests

Tests dynamic updates of the 'in2' property of the SVGFECompositeElement object

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS overComposite1.in2.baseVal is "SourceGraphic"
PASS overComposite1.in2.baseVal is "off1"
PASS successfullyParsed is true

TEST COMPLETE
